textbackend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
qdrant-client
psycopg2-binary
pydantic
fastembed